refactor: ドメインの定義
