+++ src/domain/entities/Comment.ts
@@ -1,0 +1,6 @@
+export interface Comment {
+  id: number;
+  content: string;
+  postId: number;
+  userId: number;
+}
