+++ src/domain/entities/Following.ts
@@ -1,0 +1,5 @@
+export interface Following {
+  id: number;
+  follow_id: number;
+  followed_id: number;
+}
