+++ src/domain/entities/Post.ts
@@ -1,0 +1,6 @@
+export interface Post {
+  id: number;
+  content: string;
+  likes: number;
+  userId: number;
+}
