+++ src/domain/entities/User.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: number;
+  name?: string;
+  email?: string;
+  picture?: string;
+}
